requirements.txt
treelib==1.7.0
gast==0.5.4
transformers==4.34.1
klepto
pygtrie
python_graphs==1.2.3

--find-links https://data.dgl.ai/wheels/repo.html
#dgl==0.9.1.post1
dgl-cu116==0.9.1.post1

duckdb==0.10.1
func-timeout==4.3.5
gensim==4.2.0
joblib==1.2.0
lightgbm==3.3.5
matplotlib==3.8.2
networkx==2.8.8
numpy==1.23.5
osfclient==0.0.5
pandas==1.5.3
paramiko==3.4.0
psycopg2-binary==2.9.9
pyarrow==14.0.2
scikit-learn==1.1.3
scipy==1.8.1
seaborn==0.12.2
SQLAlchemy==2.0.25
tabulate==0.9.0
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.64.1
wandb==0.16.1
scp
sqlparse==0.5.1

# deepdb requirements
git+https://github.com/SPFlow/SPFlow.git
